python_backend/requirements.txt
flask==3.1.1
requests==2.32.3
python-dotenv==1.0.1